조회수 수정(조회수 증가 시 수정시간 생성 방지) & 검색 like 기능 추가 (%검색어%)

+++ src/main/java/com/ex/memberboard/dto/BoardDetailDTO.java
@@ -24,7 +24,7 @@ public class BoardDetailDTO {
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
     private List<CommentDetailDTO> commentList;
-    private Long boardHits;
+    private int boardHits;
 
     public static BoardDetailDTO toBoardDetailDTO(BoardEntity boardEntity) {
         BoardDetailDTO boardDetailDTO = new BoardDetailDTO();
